+++ get-option.js
@@ -2,6 +2,8 @@ const defaultOptions = {
     smoothScrolling: true,
     persistentCollapse: true,
     newestItems: true,
+    updateLocation: true,
+    updateLocationDelay: 2.0,
 };
 
 function getOption(options, key) {